Add files via upload
E-commerce involves using online websites and computer tools to help businesses sell things and
interact with customers over the Internet. The E-commerce Administrator is in charge of running the
online store. On the flip side, the Shopper is the person who looks around and buys things from the
online store